requirements.txt
ops
jinja2
lightkube
lightkube-models
pydantic
validators
pytest-interface-tester